+++ cmd/edit_extended_test.go
@@ -768,112 +768,6 @@ func TestWriteAndRename_TitleChanged(t *testing.T) {
 	}
 }
 
-// --- logEditActivity tests ---
-
-func TestLogEditActivity_BasicEdit(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tk := &task.Task{ID: 1, Title: "test"}
-	logEditActivity(cfg, tk, false, "")
-
-	logPath := filepath.Join(kanbanDir, "activity.jsonl")
-	data, err := os.ReadFile(logPath) //nolint:gosec // test path
-	if err != nil {
-		t.Fatal(err)
-	}
-	got := string(data)
-	if !containsSubstring(got, "edit") {
-		t.Errorf("expected 'edit' action in log, got: %s", got)
-	}
-}
-
-func TestLogEditActivity_BlockTransition(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tk := &task.Task{ID: 1, Title: "test", Blocked: true, BlockReason: "dependency"}
-	logEditActivity(cfg, tk, false, "")
-
-	logPath := filepath.Join(kanbanDir, "activity.jsonl")
-	data, err := os.ReadFile(logPath) //nolint:gosec // test path
-	if err != nil {
-		t.Fatal(err)
-	}
-	got := string(data)
-	if !containsSubstring(got, "block") {
-		t.Errorf("expected 'block' action in log, got: %s", got)
-	}
-}
-
-func TestLogEditActivity_UnblockTransition(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tk := &task.Task{ID: 1, Title: "test", Blocked: false}
-	logEditActivity(cfg, tk, true, "")
-
-	logPath := filepath.Join(kanbanDir, "activity.jsonl")
-	data, err := os.ReadFile(logPath) //nolint:gosec // test path
-	if err != nil {
-		t.Fatal(err)
-	}
-	got := string(data)
-	if !containsSubstring(got, "unblock") {
-		t.Errorf("expected 'unblock' action in log, got: %s", got)
-	}
-}
-
-func TestLogEditActivity_ClaimTransition(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tk := &task.Task{ID: 1, Title: "test", ClaimedBy: "agent-1"}
-	logEditActivity(cfg, tk, false, "")
-
-	logPath := filepath.Join(kanbanDir, "activity.jsonl")
-	data, err := os.ReadFile(logPath) //nolint:gosec // test path
-	if err != nil {
-		t.Fatal(err)
-	}
-	got := string(data)
-	if !containsSubstring(got, "claim") {
-		t.Errorf("expected 'claim' action in log, got: %s", got)
-	}
-}
-
-func TestLogEditActivity_ReleaseTransition(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	tk := &task.Task{ID: 1, Title: "test", ClaimedBy: ""}
-	logEditActivity(cfg, tk, false, "agent-1")
-
-	logPath := filepath.Join(kanbanDir, "activity.jsonl")
-	data, err := os.ReadFile(logPath) //nolint:gosec // test path
-	if err != nil {
-		t.Fatal(err)
-	}
-	got := string(data)
-	if !containsSubstring(got, "release") {
-		t.Errorf("expected 'release' action in log, got: %s", got)
-	}
-}
 
 // --- appendBody unit tests ---
