+++ cmd/move_pick_coverage_test.go
@@ -350,135 +350,8 @@ func TestResolveTargetStatus_NextUnknownStatus(t *testing.T) {
 	}
 }
 
-// --- enforceWIPLimit ---
 
-func TestEnforceWIPLimit_NoLimit(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
 
-	err = enforceWIPLimit(cfg, "backlog", "todo")
-	if err != nil {
-		t.Errorf("expected nil with no WIP limit, got: %v", err)
-	}
-}
-
-func TestEnforceWIPLimit_Exceeded(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-	cfg.WIPLimits = map[string]int{"todo": 1}
-	if saveErr := cfg.Save(); saveErr != nil {
-		t.Fatal(saveErr)
-	}
-	createTaskFileWithStatus(t, cfg.TasksPath(), 1, "in-todo", "todo")
-
-	cfg, err = config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	err = enforceWIPLimit(cfg, "backlog", "todo")
-	if err == nil {
-		t.Fatal("expected WIP limit exceeded error")
-	}
-}
-
-// --- enforceWIPLimitForClass ---
-
-func TestEnforceWIPLimitForClass_NilClass(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Task with unknown class — ClassByName returns nil.
-	tk := &task.Task{ID: 1, Class: "nonexistent", Status: "backlog"}
-	err = enforceWIPLimitForClass(cfg, tk, "backlog", "todo")
-	if err != nil {
-		t.Errorf("expected nil when class not configured, got: %v", err)
-	}
-}
-
-func TestEnforceWIPLimitForClass_BypassColumnWIP(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Set a tight column WIP limit that would normally fail.
-	cfg.WIPLimits = map[string]int{"todo": 1}
-	if saveErr := cfg.Save(); saveErr != nil {
-		t.Fatal(saveErr)
-	}
-	createTaskFileWithStatus(t, cfg.TasksPath(), 2, "existing-todo", "todo")
-
-	cfg, err = config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Expedite class bypasses column WIP by default.
-	tk := &task.Task{ID: 1, Class: "expedite", Status: "backlog"}
-	err = enforceWIPLimitForClass(cfg, tk, "backlog", "todo")
-	if err != nil {
-		t.Errorf("expected nil when class bypasses column WIP, got: %v", err)
-	}
-}
-
-func TestEnforceWIPLimitForClass_ClassWIPExceeded(t *testing.T) {
-	kanbanDir := setupBoard(t)
-	cfg, err := config.Load(kanbanDir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create an expedite task (default WIP limit = 1).
-	writeTaskWithClass(t, cfg.TasksPath(), 2, "other-expedite", "todo", "expedite")
-
-	tk := &task.Task{ID: 3, Class: "expedite", Status: "backlog"}
-	err = enforceWIPLimitForClass(cfg, tk, "backlog", "todo")
-	if err == nil {
-		t.Fatal("expected class WIP exceeded error")
-	}
-	var cliErr *clierr.Error
-	if !errors.As(err, &cliErr) {
-		t.Fatalf("expected clierr.Error, got %T: %v", err, err)
-	}
-	if cliErr.Code != clierr.ClassWIPExceeded {
-		t.Errorf("code = %q, want %q", cliErr.Code, clierr.ClassWIPExceeded)
-	}
-}
-
-// --- countByClass ---
-
-func TestCountByClass_Counts(t *testing.T) {
-	tasks := []*task.Task{
-		{ID: 1, Class: "expedite"},
-		{ID: 2, Class: "expedite"},
-		{ID: 3, Class: "standard"},
-	}
-	// Exclude ID 1 — should only count ID 2.
-	got := countByClass(tasks, "expedite", 1)
-	if got != 1 {
-		t.Errorf("countByClass = %d, want 1", got)
-	}
-}
-
-func TestCountByClass_NoneMatch(t *testing.T) {
-	tasks := []*task.Task{
-		{ID: 1, Class: "standard"},
-	}
-	got := countByClass(tasks, "expedite", 0)
-	if got != 0 {
-		t.Errorf("countByClass = %d, want 0", got)
-	}
-}
 
 // --- outputMoveResult ---
